requirements.txt
discord.py[voice]
PyNaCl
python-dev
gtts
youtube_dl